HttpStore: skip test testAddAndDeleteStatementsOnDefaultGraph #36
Because we dont know if the target server supports
write access to default graph and it may throw an exception,
if not. because of that we just dont test it to avoid confusion.

It is not clear, what the target server supports, so you have
to care while you using it, except you know what is behind the
endpoint. than you can adapt your queries, ...

Further adaption on AbstractSparqlStore to let pass addStatements
calls with no graph information given.